perf(core): native uint8 pixel path for 8-bit HEVC (+15% 4K, +4% 1080p)
Store and process 8-bit HEVC content at one byte per sample (uint8) instead of
uint16, halving picture memory and memory traffic through motion compensation,
SAO, and deblocking. 10-bit content stays uint16. Decode is bit-identical either
way — all 106 conformance/oracle tests pass (8- and 10-bit), AddressSanitizer-
clean, and the new SIMD kernels were adversarially verified exhaustively.

Storage seam + dispatch:
- Picture::allocate flips bytes_per_sample to 1 when every plane is <=8-bit.
- The pixel-touching filters are templated on the plane Sample type and dispatch
  once per picture on bytes_per_sample: apply_deblocking / deblock_luma_edge and
  apply_sao join the already-templated reconstruct / intra / interpolation paths.
- sao_backup_ is now byte-sized (uint8 vector, sized in bytes) so it backs either
  storage width.

Native uint8 SIMD (the win — without it the flip regresses ~16-33% by losing the
uint16 SIMD):
- Motion comp: simd_filter_row_u8 / simd_copy_row_u8 load bytes and widen to
  int16, reusing the madd kernel (bit-exact, half the load bytes).
- SAO: true 16-lane epu8 edge/band kernels (sao_eo_row_u8 / sao_bo_row_u8) —
  2x the throughput of the uint16 path. The dominant 4K win.
- Deblocking: dbk_load4 / dbk_store4 / deblock_luma_hor_simd templated on Sample
  (uint8 widens to int32, narrows back with packus).

Output ABI (deletes the deployment's planeTo8 pass for 8-bit):
- HEVCFrame gains a trailing bytes_per_sample field (existing offsets unchanged).
- get_frame / get_drained_frame, write_yuv, and the CLI multi-frame writer read
  native u8 storage directly.
- JS: HEVCFrame/HEVCFrameView planes are Uint8Array|Uint16Array + bytesPerSample;
  drainViews()/copyPlane/_readFrameView branch on it (HEAPU8 for 8-bit, no >>1),
  so drainViews() hands 8-bit content native u8 heap views — zero-conversion to
  a texture. tools/wasm_hash.mjs and the uint16-assuming unit tests updated to be
  storage-aware.

WASM single-thread decode vs the prior uint16+SIMD path: bbb4k 12.4 -> 14.2 fps
(+15%), bbb1080 54.2 -> 56.5 fps (+4%).

+++ src/decoding/decoder.h
@@ -89,8 +89,10 @@ private:
     std::vector<uint8_t> edge_flags_h_buf_;
     std::vector<DecodingContext::SaoParams> sao_params_buf_;
 
-    // Phase 9: SAO backup buffers (reused across frames to avoid per-frame allocation)
-    std::vector<uint16_t> sao_backup_[3];
+    // Phase 9: SAO backup buffers (reused across frames to avoid per-frame
+    // allocation). Byte-sized so they hold either uint8 (native 8-bit) or
+    // uint16 plane bytes; apply_sao reinterprets per the plane Sample type.
+    std::vector<uint8_t> sao_backup_[3];
 
     // Phase 10: slice index per CTU
     std::vector<uint8_t> slice_idx_buf_;
